Refactor requests classes for reusability across use cases
